+++ FindPanel/OgreTextFindComponents/OgreTextFindComponentEnumerator.h
@@ -23,13 +23,13 @@
 #else
     unsigned			*_indexes, _count;
 #endif
-	int					_nextIndex;
-    int					_terminalIndex;
+	NSInteger			_nextIndex;
+    NSInteger			_terminalIndex;
     BOOL				_inSelection;
 }
 
 - (id)initWithBranch:(OgreTextFindBranch*)aBranch inSelection:(BOOL)inSelection;
-- (void)setTerminalIndex:(int)index;
-- (void)setStartIndex:(int)index;
+- (void)setTerminalIndex:(NSInteger)index;
+- (void)setStartIndex:(NSInteger)index;
 
 @end
