+++ RegularExpression/OGRegularExpressionEnumeratorPrivate.h
@@ -16,7 +16,7 @@
 #import <OgreKit/OGString.h>
 
 // aUTF16StringのUTF16文字長
-static inline unsigned Ogre_UTF16charlen(unichar *const aUTF16String)
+static inline NSUInteger Ogre_UTF16charlen(unichar *const aUTF16String)
 {
 	unichar UTF16Char = *aUTF16String;
 	
@@ -30,7 +30,7 @@ static inline unsigned Ogre_UTF16charlen(unichar *const aUTF16String)
 }
 
 // aUTF16Stringより１文字前のUTF16文字長
-static inline unsigned Ogre_UTF16prevcharlen(unichar *const aUTF16String)
+static inline NSUInteger Ogre_UTF16prevcharlen(unichar *const aUTF16String)
 {
     unichar UTF16Char = *(aUTF16String - 1);
 	if ((UTF16Char <= 0x9FFF) || (UTF16Char >= 0xE000)) return 1;       // 1 code point
@@ -58,7 +58,7 @@ static inline unsigned Ogre_UTF16prevcharlen(unichar *const aUTF16String)
 /*********************
  * private accessors *
  *********************/
-- (void)_setTerminalOfLastMatch:(int)location;
+- (void)_setTerminalOfLastMatch:(NSInteger)location;
 - (void)_setIsLastMatchEmpty:(BOOL)yesOrNo;
 - (void)_setStartLocation:(NSUInteger)location;
 - (void)_setNumberOfMatches:(NSUInteger)aNumber;
